+++ CHANGELOG.md
@@ -13,7 +13,8 @@ Recommendation: for ease of reading, use the following order:
 
 ## [Unreleased]
 ### Changed
-- GraphQL: accountListFlows returns list sorted by status and last event time
+- GraphQL: flows are listed ordered by status and last event time
+- Merged two methods(`saveEnvVariable` and `modifyEnvVariable`) from `DatasetEnvVarsMut` info one `upsertEnvVariable`
 ### Fixed
 - GQL api flows queries now fetch dataset polling source only once per dataset(and only if Ingest flow type is here)
 - Flow trigger status now become disable on flow fail